backend/pyproject.toml
[project]
name = "flow-backend"
version = "0.1.0"
description = "Flow - AI Knowledge Tree Generator"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0",
    "python-multipart>=0.0.12",
    "openai>=1.0.0",
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "PyMuPDF>=1.24.0",
    "python-dotenv>=1.0.0",
]
